fix warnings
fix warnings

+++ src/fragments/34/fragment34.h
@@ -453,7 +453,7 @@ void func_81404024(u8* texture, s32 width, s32 height);
 void func_81404094(u8* texture, s32 width, s32 height);
 void func_814040E0(u8* texture, s32 width, s32 height);
 void func_81404150(u8* texture, s32 width, s32 height);
-void func_8140419C(u8* texture, s32 width, s32 height);
+void func_8140419C(u32* texture, s32 width, s32 height);
 void func_814041BC(u8* texture, s32 width, s32 height);
 void func_814041DC(u8* texture, s32 width, s32 height);
 void func_81404228(u8* texture, s32 width, s32 height);
